+++ apps/ledger-live-desktop/src/renderer/screens/accounts/AccountRowItem/Placeholder.tsx
@@ -5,6 +5,7 @@ import { ModularDrawerLocation } from "LLD/features/ModularDrawer/enums";
 import { useOpenAssetFlow } from "LLD/features/ModularDrawer/hooks/useOpenAssetFlow";
 import Box from "~/renderer/components/Box";
 import IconPlus from "~/renderer/icons/Plus";
+import { MAD_SOURCE_PAGES } from "LLD/features/ModularDrawer/analytics/types";
 
 const AddAccountButton = styled(Box)`
   border: 1px dashed rgba(153, 153, 153, 0.3);
@@ -27,8 +28,10 @@ const AddAccountButton = styled(Box)`
 `;
 const Placeholder = () => {
   const { t } = useTranslation();
-  const { openAssetFlow } = useOpenAssetFlow(ModularDrawerLocation.ADD_ACCOUNT);
-
+  const { openAssetFlow } = useOpenAssetFlow(
+    ModularDrawerLocation.ADD_ACCOUNT,
+    MAD_SOURCE_PAGES.ACCOUNTS_PAGE,
+  );
   const handleAddAccountClick = () => {
     openAssetFlow(true);
   };