+++ libs/ledger-services/trust/CHANGELOG.md
@@ -1,12 +1,25 @@
 # @ledgerhq/ledger-trust-service
 
-## 0.3.4-nightly.0
+## 0.3.4
 
 ### Patch Changes
 
-- Updated dependencies [[`8551c28`](https://github.com/LedgerHQ/ledger-live/commit/8551c280f24f7bd4475c6cc12f1b1d92636d9357)]:
-  - @ledgerhq/live-env@2.9.1-nightly.0
-  - @ledgerhq/live-network@2.0.9-nightly.0
+- Updated dependencies []:
+  - @ledgerhq/live-network@2.0.9
+
+## 0.3.4-next.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/live-network@2.0.9-next.1
+
+## 0.3.4-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/live-network@2.0.9-next.0
 
 ## 0.3.3