+++ libs/ledgerjs/packages/hw-app-multiversx/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ledgerhq/hw-app-elrond
 
+## 6.24.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.6
+
+## 6.24.2-next.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.6-next.1
+
+## 6.24.2-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.6-next.0
+
 ## 6.24.1
 
 ### Patch Changes