+++ libs/ledgerjs/packages/hw-app-polkadot/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.34.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.6
+
+## 6.34.2-next.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.6-next.1
+
+## 6.34.2-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.6-next.0
+
 ## 6.34.1
 
 ### Patch Changes